### 2025-02-06 #7757 @editorialbot
[PRE REVIEW]: Segmenteverygrain: A Python module for segmentation of grains in a variety of image types
**Submitting author:** @zsylvester (<a href="http://orcid.org/0000-0002-4890-4063">Zoltan Sylvester</a>)
**Repository:** https://github.com/zsylvester/segmenteverygrain
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
