+++ paper/paper.md
@@ -46,22 +46,22 @@ bibliography: paper.bib
 
 # Summary
 
-`Segmenteverygrain` is a Python module that addresses the need for quantifying grain size in granular materials. It combines the `Segment Anything Model` (SAM) with a U-Net-style convolutional neural network to detect and measure grains or clasts in diverse image types, ranging from photomicrographs of sand, mineral grains, and thin sections to photographic images of gravel and boulder fields. Compared to similar tools such as `GrainSight` [@Azzam:2024] and `ImageGrains` [@Mair:2022; @Mair:2024], `Segmenteverygrain` offers greater flexibility and robustness, though it requires basic familiarity with Python and Jupyter notebooks. `Segmenteverygrain` supports segmentation of large images, georeferencing, and interactive editing of the results. The combined U-Net-SAM approach provides high-quality segmentation results, allows for fine-tuning with minimal training data, and can be expanded to images with other object types.
+`Segmenteverygrain` is a Python module that addresses the need for quantifying grain size in granular materials. It combines the `Segment Anything Model` (SAM) with a U-Net-style convolutional neural network to detect and measure grains or clasts in diverse image types, ranging from photomicrographs of sand, mineral grains, and thin sections to photographic images of gravel and boulder fields. `Segmenteverygrain` supports segmentation of large images, georeferencing, and interactive editing of the results.
 
 # Statement of need
 
 Grain morphology, in particular size and shape, are key parameters that influence the physical and chemical properties of granular materials. Quantitative estimates of these parameters  are important in a broad range of fields, including:
 
-* geomorphology, sedimentology, stratigraphy, and paleontology (sediment transport and morphodynamics);
-* subsurface reservoir quality and engineering (compaction and cementation, hydrocarbon production, aquifer management, CO2 storage);
-* structural geology, petrology and geochemistry (grain-size sensitive rheology, nucleation and cooling rates of melts, chemical reactiveness and alteration);
-* civil and geotechnical engineering (foundation design, erosion control, sedimentation);
-* environmental science (pollutant transport, absorption potential, biochemical rock-water interaction);
-* materials science (gravel and sand for construction; manufacturing of glass, concrete, and asphalt).
+* geomorphology, sedimentology, stratigraphy, and paleontology;
+* subsurface reservoir quality and engineering;
+* structural geology, petrology and geochemistry;
+* civil and geotechnical engineering;
+* environmental science;
+* materials science.
 
-While the grain size of loose sand and finer-than-sand sediment can be measured accurately on bulk sample using laser particle size analyzers [@Blott:2004], quantifying the grain size of small samples, cemented rock or of sediment that is coarser than medium sand commonly requires manual measurement of hundreds of individual grain lengths and widths. This is increasingly done via the analysis of digital images, often in a grain-by-grain fashion. In recent years, however, numerous studies have illustrated the promise of automated image analysis and/or machine learning (ML) approaches [@Buscombe:2020; @Tang:2020; @Mair:2022; @Chen:2023; @Prieur:2023; @Mair:2024; @Azzam:2024]. While these studies clearly show that ML techniques are superior to both painstaking manual data collection and conventional image processing techniques [e.g., @Purinton:2021], they tend to focus on a relatively narrow range of image types, e.g., gravel on fluvial bars [@Mair:2022; @Mair:2024], boulder fields on planetary surfaces [@Prieur:2023; @Robin:2024], or petrographic images [@Tang:2020; @Azzam:2024].
[…]
