+++ paper/paper.md
@@ -50,7 +50,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Grain morphology, in particular size and shape, are key parameters that influence the physical and chemical properties of granular materials. Quantitative estimates of these parameters  are important in a broad range of fields, including:
+Grain size and shape are key parameters that influence the physical and chemical properties of granular materials. Quantitative estimates of these parameters are important in a numnber of fields, including:
 
 * geomorphology, sedimentology, stratigraphy, and paleontology;
 * subsurface reservoir quality and engineering;
@@ -61,7 +61,7 @@ Grain morphology, in particular size and shape, are key parameters that influenc
 
 While the grain size of loose sand and finer-than-sand sediment can be measured accurately on bulk sample using laser particle size analyzers [@Blott:2004], quantifying the grain size of small samples, cemented rock or of sediment that is coarser than medium sand commonly requires manual measurement of hundreds of individual grain lengths and widths. In recent years, numerous studies have illustrated the promise of automated image analysis and/or machine learning (ML) approaches [@Buscombe:2020; @Tang:2020; @Mair:2022; @Chen:2023; @Prieur:2023; @Mair:2024; @Azzam:2024]. While these studies clearly show that ML techniques are superior to both  manual data collection and conventional image processing techniques [e.g., @Purinton:2021], they tend to focus on a narrow range of image types, e.g., gravel on fluvial bars [@Mair:2022; @Mair:2024], boulder fields on planetary surfaces [@Prieur:2023; @Robin:2024], or petrographic images [@Tang:2020; @Azzam:2024].
 
-With the emergence of large image segmentation models that have been trained on millions of images [e.g., @Kirillov:2023; @Ravi:2024], the opportunity arises for the utilization of these models to detect a variety of grains and grain-like objects in a broad range of image types. `Segmenteverygrain` is a Python module that takes advantage of the Segment Anything Model (SAM) [@Kirillov:2023] to generate accurate segmentation masks for grains in any image, as long as the grains are relatively well defined. To automate this process, `Segmenteverygrain` uses a U-Net-style convolutional neural network to create prompts (pixel coordinates of grain centers) for SAM; and ensures that the resulting masks contain no duplicates and do not overlap. In general, SAM masks are more robust and accurate than the U-Net output (\autoref{fig:1}). The U-Net model uses patches as input and output; to reduce edge effects, Hann-window-based weighting is used on overlapping patches [@Pielawski:2020]. The U-Net model has been trained on 66 images of a variety of grains that were split into 44,533 patches of 256x256 pixels. Some of these images were labeled by manually outlining individual grains; for others, we used SAM in interactive mode by clicking on each object to provide a prompt. Segmenting every grain eliminates the problem of selecting a representative sample of grains - as long as the image itself is representative.
+With the emergence of large image segmentation models that have been trained on millions of images [e.g., @Kirillov:2023; @Ravi:2024], the opportunity arises for the utilization of these models to detect a variety of grains and grain-like objects in a broad range of image types. `Segmenteverygrain` is a Python module that takes advantage of the Segment Anything Model (SAM) [@Kirillov:2023] and uses a U-Net-style convolutional neural network to create prompts (pixel coordinates of grain centers) for SAM. It ensures that the resulting masks contain no duplicates and do not overlap. In general, SAM masks are more robust and accurate than the U-Net output (\autoref{fig:1}). The U-Net model uses patches as input and output; to reduce edge effects, Hann-window-based weighting is used on overlapping patches [@Pielawski:2020]. The U-Net model has been trained on 66 images of a variety of grains that were split into 44,533 patches of 256x256 pixels.
 
[…]
