+++ paper/paper.md
@@ -76,32 +76,24 @@ So far, `Segmenteverygrain` has been successfully used on:
  
 ![Photomicrograph (plane polarized light) of a sandstone in thin section (A), output of the \`Segmenteverygrain\` segmentation (B), size distributions of the major and minor grain axes (C), and the area-weighted size distributions (D). Photomicrograph from \[@Prodanovic:2019\]. Major grain axis lengths are shown in blue, minor grain axis lengths in orange.\label{fig:3}](joss_paper_fig_3.jpg)
 
-`Segmenteverygrain` also adds functionality for segmenting large images (with a size limited by computer memory), working with georeferenced data, and interactively deleting, merging, and adding grains to the segmentation result.
-
 # Key functionality
 
-The base U-Net model that is available in the `Segmenteverygrain` repository works relatively well on a variety of image types. However, it is recommended that it is first tested on a small image (e.g., 2000 x 300 pixels), to see how well the U-Net model captures the difference between the grains and the background.
+The U-Net model that is available in the `Segmenteverygrain` repository works relatively well on a variety of image types. However, it is recommended that it is first tested on a small image (e.g., 2000 x 3000 pixels), to see how well the U-Net model captures the difference between the grains and the background. If additional fine tuning is necessary, `Segmenteverygrain` has tools for interactively deleting, merging, and adding grains to generate training data. The training data can be used to fine tune the U-Net model.
 
-Looking at the U-Net output, if it is obvious that there is room for improvement, a few small images should be used to derive training data for fine tuning the U-Net model. This can be done either by first running the segmentation workflow and then cleaning up the result, or, if the output is of low quality, by using SAM to do the segmentation. `Segmenteverygrain` makes this process straightforward as it has tools for interactively deleting, merging, and adding grains.
+The `predict_large_image` function can be used to run the segmentation of larger images that contain thousands or tens of thousands of grains. This is done by running the U-Net + SAM predictions on smaller tiles of the input image (default size of 2000 x 2000 pixels), and collecting the grains into a list without duplicates.
 
-Next, the training data (pairs of images and segmentation masks) can be used to fine tune the U-Net model. The `patchify_training_data` function creates patches of 256 x 256 pixels, which are split into training, validation, and test sets using the `create_train_val_test_data` function. Data augmentation is applied during training (`load_and_preprocess` function). The `create_and_train_model` function loads the weights of the base model, runs the training, and evaluates the new model.
-
-Once the `Segmenteverygrain` workflow has been tested on smaller images, the `predict_large_image` function can be used to run the segmentation of larger images that contain thousands or tens of thousands of grains.
-
-This is done by running the U-Net + SAM predictions on smaller tiles of the input image (default size of 2000 x 2000 pixels), and collecting the grains into a list without duplicates.
-
[…]
