+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Segmenteverygrain: A Python module for segmentation of grains in a variety of image types'
+title: 'Segmenteverygrain: A Python module for grain segmentation'
 tags:
   - Python
   - geology
@@ -61,28 +61,28 @@ Grain size and shape are key parameters that influence the physical and chemical
 
 In recent years, numerous studies have illustrated the promise of automated image analysis and/or machine learning (ML) approaches [@Buscombe:2020; @Tang:2020; @Mair:2022; @Chen:2023; @Prieur:2023; @Mair:2024; @Azzam:2024]. While these studies clearly show that ML techniques are superior to both  manual data collection and conventional image processing techniques [e.g., @Purinton:2021], they tend to focus on a narrow range of image types, e.g., gravel on fluvial bars [@Mair:2022; @Mair:2024], boulder fields on planetary surfaces [@Prieur:2023; @Robin:2024], or petrographic images [@Tang:2020; @Azzam:2024].
 
-With the emergence of large image segmentation models that have been trained on millions of images [e.g., @Kirillov:2023; @Ravi:2024], the opportunity arises for the utilization of these models to detect a variety of grains and grain-like objects in a broad range of image types. `Segmenteverygrain` is a Python module that takes advantage of the Segment Anything Model (SAM) [@Kirillov:2023] and uses a U-Net-style convolutional neural network to create prompts (pixel coordinates of grain centers) for SAM. It ensures that the resulting masks contain no duplicates and do not overlap. In general, SAM masks are more robust and accurate than the U-Net output (\autoref{fig:1}). The U-Net model uses patches as input and output; to reduce edge effects, Hann-window-based weighting is used on overlapping patches [@Pielawski:2020]. The U-Net model has been trained on 66 images of a variety of grains that were split into 44,533 patches of 256x256 pixels.
+With the emergence of large image segmentation models that were trained on millions of images [e.g., @Kirillov:2023; @Ravi:2024], the opportunity arises to use these models to detect a variety of grains in a broad range of image types. `Segmenteverygrain` is a Python module that takes advantage of the Segment Anything Model (SAM) [@Kirillov:2023] and uses a U-Net-style convolutional neural network to create prompts for SAM. It ensures that the resulting masks contain no duplicates and do not overlap. In general, SAM masks are more robust and accurate than the U-Net output (\autoref{fig:1}). The U-Net model uses patches as input and output; to reduce edge effects, Hann-window-based weighting is used on overlapping patches [@Pielawski:2020]. The U-Net model was trained on 66 images of a variety of grains that were split into 44,533 patches of 256x256 pixels.
 
-![Photo of fluvial gravel (A), output of the U-Net segmentation (B), and the result of the SAM segmentation (C). Grains 1-5 are all detected correctly, despite the fact that the grain boundary class in the U-Net segmentation does not fully delineate them. Patch #6 in the final segmentation result is a false positive that can be deleted manually. \label{fig:U-Net_sam_comparison}\label{fig:1}](joss_paper_fig_1.jpg)
+![Photo of fluvial gravel (A), output of the U-Net segmentation (B), and the result of the SAM segmentation (C). \label{fig:U-Net_sam_comparison}\label{fig:1}](joss_paper_fig_1.jpg)
 
-So far, `Segmenteverygrain` has been successfully used on:
+`Segmenteverygrain` has been successfully used on:
 
 * images of boulder fields on asteroids [@Robin:2024];
 * photographs of gravel and cobbles on beaches [@Roberts:2024] and on fluvial bars (\autoref{fig:2});  
 * photomicrographs of thin sections of sandstones (\autoref{fig:3}) and oolitic limestones;  
[…]
