+++ paper/paper.bib
@@ -16,23 +16,6 @@
 	pages = {243--254},
 }
 
-@article{Blott:2004,
-	title = {Particle size analysis by laser diffraction},
-	volume = {232},
-	copyright = {https://doi.org/10.15223/policy-002},
-	issn = {0305-8719, 2041-4927},
-	url = {https://www.lyellcollection.org/doi/10.1144/GSL.SP.2004.232.01.08},
-	doi = {10.1144/GSL.SP.2004.232.01.08},
-	language = {en},
-	number = {1},
-	urldate = {2024-12-16},
-	journal = {Geological Society, London, Special Publications},
-	author = {Blott, Simon J. and Croft, Debra J. and Pye, Kenneth and Saye, Samantha E. and Wilson, Helen E.},
-	month = jan,
-	year = {2004},
-	pages = {63--73},
-}
-
 @article{Buscombe:2013,
 	title = {Transferable wavelet method for grain-size distribution from images of sediment surfaces and thin sections, and other natural granular patterns},
 	volume = {60},
@@ -404,19 +387,19 @@
 	pages = {e453},
 }
 
-@software{Gillies:2019,
-	author =    {Sean Gillies and others},
+@software{Gillies:2013,
+	author =    {Sean Gillies et al.},
 	organization = {Mapbox},
 	title =     {Rasterio: geospatial raster I/O for {Python} programmers},
-	year =      {2013--},
+	year =      {2013},
 	url = "https://github.com/rasterio/rasterio"
 }
 
 @Misc{Gillies:2007,
-	author = {Sean Gillies and others},
+	author = {Sean Gillies et al.},
 	organization = {toblerity.org},
 	title = {Shapely: manipulation and analysis of geometric objects},
-	year =  {2007--},
+	year =  {2007},
 	url = "https://github.com/Toblerity/Shapely"
 }
 
