### 2025-03-20 #7757 @kthyng
Hi @anastassiavybornova! I would say start at step 4 to see what sorts of checks to be thinking about for submissions. I have already done checks, but you'll want to feel confident about the submission details too.

> as editor i can @ mention the reviewers suggested by the author?

Yes, once you've checked the reviewers suggested to see if they seem like a reasonable fit, have an at least somewhat active github record, and check in the reviewer database to make sure they aren't already reviewing something, you can go ahead and ping 1 or a few reviewers to see if they're available.
