### 2025-03-21 c9767bb
add predict_from_prompts() to handle arbitrary type/number of prompts
