+++ segmenteverygrain/segmenteverygrain.py
@@ -363,6 +363,47 @@ def two_point_prompt(x1, y1, x2, y2, image, predictor, ax=False):
         ax.fill(sx, sy, facecolor=color, edgecolor='k', alpha=0.5)
     return sx, sy
 
+def predict_from_prompts(predictor, box=None, points=None, point_labels=None):
+    """
+    Perform a point-prompt-based segmentation using the SAM model. 
+
+    Parameters
+    ----------
+        predictor
+            SAM predictor
+        image: numpy.ndarray
+            Input image
+        box: list, optional
+            Selection box coordinates, as [xmin, ymin, xmax, ymax]
+        points: list, optional
+            Prompt coordinates, as [(x1, y1), (x2, y2)...]
+        point_labels: list, optional
+            Bool values for each point, True if inside the object
+
+    Returns
+    -------
+        contour: list
+            Points along resulting contour, as [(x1, y1), (x2, y2)...]
+    """
+    # Format input
+    if points is not None:
+        points = np.array(points)
+        point_labels = np.array(point_labels)
+    if box is not None:
+        box = np.array(box)[None, :]
+    # Do segmentation
+    masks, _, _ = predictor.predict(
+        point_coords=points,
+        point_labels=point_labels,
+        box=box,
+        # TODO: Is multimask_output=True better when using only one prompt?
+        multimask_output=False
+    )
+    # Find and return points along contour of detected object
+    contours = measure.find_contours(masks[0], 0.5)
+    sx, sy = contours[0][:,1], contours[0][:,0]
+    return sx, sy
+
 def find_overlapping_polygons(polygons, min_overlap=0.4):
     """
     Finds and returns a list of overlapping polygons from the given list of polygons using spatial indexing.
