### 2025-03-21 9f40f15
add interactive interface in python submodule
