+++ segmenteverygrain/interactions/interactions.py
@@ -1,0 +1,1255 @@
+import keras.utils
+import logging
+import matplotlib as mpl
+import matplotlib.style as mplstyle
+import matplotlib.patches as mpatches
+import matplotlib.pyplot as plt
+import matplotlib.widgets as mwidgets
+import numpy as np
+import pandas as pd
+import rasterio.features
+import segmenteverygrain
+import shapely
+import skimage
+from tqdm import tqdm
+
+# Images larger than this will be downscaled
+# 4k resolution is (2160, 4096)
+IMAGE_MAX_SIZE = np.asarray((2160, 4096))
+
+# Init logger
+logger = logging.getLogger(__name__)
+
+# Speed up rendering a little?
+mplstyle.use('fast')
+
+# HACK: Bypass large image restriction
+from PIL import Image
+Image.MAX_IMAGE_PIXELS = None
+
+# HACK: Don't reset zoom level when pressing 'c' to create a grain
+if 'c' in mpl.rcParams['keymap.back']:
+    mpl.rcParams['keymap.back'].remove('c')
+
+# HACK: Attach parent grain reference to the mpatches.Polygon class
+# Makes it easy to use matplotlib "pick" event when clicking on a grain
+mpatches.Polygon.grain = None
+
+
+class Grain(object):
+    ''' Stores data and plot representation for a single grain. '''
+        
+    def __init__(self, xy: np.ndarray, data: pd.Series=None):
+        '''
+        Parameters
+        ----------
+        xy : list of (x, y) tuples
+            Coordinates to draw this grain as a polygon.
+        data : pd.Series (optional)
+            Row from a DataFrame containing information about this grain. 
+        '''
+        
+        # Input
+        self.data = data
+        self.xy = np.array(xy)
+        
+        # Grain properties to calculate (skimage)
+        # {name: dimensionality}
+        self.region_props = {
+            'area': 2,
+            'centroid': 0,
+            'major_axis_length': 1,
+            'minor_axis_length': 1, 
+            'orientation': 0,
+            'perimeter': 1,
+            'max_intensity': 0,
+            'mean_intensity': 0,
+            'min_intensity': 0
+        }
+        
+        # Display
+        self.default_props = {
+            'alpha': 0.6
+            # facecolor is set when patch is created
+        }
+        self.selected_props = {
+            'alpha': 1.0,
+            'facecolor': 'lime'
+        }
+        self.patch = None
+        self.selected = False
+
+    @property
+    def polygon(self) -> shapely.Polygon:
+        ''' 
+        Return a shapely.Polygon representing the matplotlib patch.
+        
+        Returns
+        -------
+        poly : shapely.Polygon
+            Polygon representing the boundaries of this grain.
+        '''
+        poly = shapely.Polygon(self.xy.T)
+        return poly
+
+    def measure(self, image: np.ndarray) -> pd.Series:
+        '''
+        Calculate grain information from image and matplotlib patch.
+        Overwrites self.data.
+
+        Parameters
+        ----------
+        image : np.ndarray
+            Background image, used to calculate region intensity in self.data.
+
+        Returns
+        -------
+        self.data : pd.Series
+            Row for a DataFrame containing computed grain info.
+        '''
+        # Get rasterized shape
+        # TODO: Just use a patch of the image and then convert coords
+        rasterized = rasterio.features.rasterize(
+            [self.polygon], out_shape=image.shape[:2])
+        # Calculate region properties
+        data = pd.DataFrame(skimage.measure.regionprops_table(rasterized,
+            intensity_image=image, properties=self.region_props.keys()))
+        if len(data):
+            self.data = data.iloc[0]
+        else:
+            # TODO: Diagnose why this happens sometimes
[…]
