### 2025-03-22 #7757 @anastassiavybornova
Hello again! I've (re-)checked that all the boxes below can indeed be ticked :)

**Software**
- [x] $\exists$ software description
- [x] $\exists$ OSI-approved license
- [x] $\exists$ installation instructions
- [x] $\exists$ documentation
- [x] $\exists$ tests

- [x] Paper compiles
- [x] Paper length 250-1000 words

**Application**
- [x] Substantial scholarly effort

And will now proceed to contact the potential reviewers.
