### 2025-03-22 #7757 @anastassiavybornova
@elliesch @nialov @jessepisel @kwinkunks @pr4deepr would you be up for reviewing this JOSS submission?

> [View article proof on GitHub](https://github.com/openjournals/joss-papers/blob/joss.07757/joss.07757/10.21105.joss.07757.pdf)
