### 2025-04-18 zsylvester/segmenteverygrain#23 @jessepisel
Georeferencing notebook example image location
This is so much fun to work with 🤓! I am running through the notebooks to [assess the functionality](https://github.com/openjournals/joss-reviews/issues/7953#issuecomment-2770653683) and cannot find the image for [cell 3 in this notebook](https://github.com/zsylvester/segmenteverygrain/blob/main/segmenteverygrain/Segment_every_grain_w_georeferencing.ipynb).

It would be great to have all the example images for the notebooks organized into one location so that new users can pick and choose which ones they want to experiment when they are getting up and running with the package.
