### 2025-04-18 e744a10
Update README.md
add contributing and reporting issues section
