### 2025-04-18 3df2f2a
add interactions submodule, defining new interactive interface
