+++ segmenteverygrain/interactions/__init__.py
@@ -1,0 +1,1 @@
+from .interactions import *
