+++ segmenteverygrain/interactions/interactions.py
@@ -1,0 +1,1502 @@
+# Standard imports
+import copy
+# Pip imports
+from PIL import Image
+import keras.utils
+import logging
+import matplotlib as mpl
+import matplotlib.style as mplstyle
+import matplotlib.patches as mpatches
+import matplotlib.pyplot as plt
+import matplotlib.widgets as mwidgets
+import numpy as np
+import pandas as pd
+import rasterio.features
+import segment_anything
+import shapely
+import skimage
+import skimage.measure
+from tqdm import tqdm
+# Local imports
+import segmenteverygrain
+
+# Images larger than this will be downscaled
+# 4k resolution is (2160, 4096)
+IMAGE_MAX_SIZE = np.asarray((2160, 4096))
+
+# Init logger
+logger = logging.getLogger(__name__)
+
+# Speed up rendering a little?
+mplstyle.use('fast')
+
+# HACK: Bypass large image restriction
+Image.MAX_IMAGE_PIXELS = None
+
+# HACK: Don't reset zoom level when pressing 'c' to create a grain
+if 'c' in mpl.rcParams['keymap.back']:
+    mpl.rcParams['keymap.back'].remove('c')
+
+# HACK: Attach parent grain reference to the mpatches.Polygon class
+# Makes it easy to use matplotlib "pick" event when clicking on a grain
+mpatches.Polygon.grain = None
+
+
+class Grain(object):
+    ''' Stores data and plot representation for a single grain. '''
+
+    def __deepcopy__(self, memo: dict):
+        ''' Custom deepcopy. Avoids copying the reference image. '''
+        # Create a new Grain instance
+        new_grain = self.__class__(self.xy.copy())
+        # Copy all attributes except for the image
+        for k, v in self.__dict__.items():
+            if k != 'image':
+                setattr(new_grain, k, copy.deepcopy(v, memo))
+        return new_grain
+
+    def __init__(self, xy: np.ndarray, image: np.ndarray = None):
+        '''
+        Parameters
+        ----------
+        xy : list of (x, y) tuples
+            Coordinates to draw this grain as a polygon.
+        image : np.ndarray (optional)
+            Image in which grain was detected. Used to measure color info.
+        '''
+
+        # Input
+        self.image = image
+        self.xy = np.array(xy)
+
+        # Metrics
+        self.data = None
+        # Metrics to calculate {name: dimensionality, for unit conversion}
+        self.region_props = {
+            'area': 2,
+            'centroid': 0,
+            'major_axis_length': 1,
+            'minor_axis_length': 1,
+            'orientation': 0,
+            'perimeter': 1,
+            'max_intensity': 0,
+            'mean_intensity': 0,
+            'min_intensity': 0
+        }
+
+        # Display
+        self.default_props = {
+            'alpha': 0.6
+            # facecolor is set when patch is created
+        }
+        self.selected_props = {
+            'alpha': 1.0,
+            'facecolor': 'lime'
+        }
+        self.axes = []
+        self.patch = None
+        self.selected = False
+
+    @property
+    def polygon(self) -> shapely.Polygon:
+        ''' 
+        Return a shapely.Polygon representing the matplotlib patch.
+
+        Returns
+        -------
+        shapely.Polygon
+            Polygon representing the boundaries of this grain.
+        '''
+        return shapely.Polygon(self.xy.T)
+
+    def measure(self, raster: bool = False) -> pd.Series:
+        '''
+        Calculate grain information from image and matplotlib patch.
+        Overwrites self.data.
+
+        Parameters
+        ----------
+        raster : bool
+            Whether to use the "old" method of measuring grain properties.
+            If True, will measure all properties from a raster representation.
+            If False, will prefer measurements directly from grain coords.
+
+        Returns
[…]
