### 2025-04-20 4e7650e
measure grain when drawing patch; ensures grain is measured when created from plot
