### 2025-04-28 zsylvester/segmenteverygrain#23 @zsylvester
Thanks @jessepisel for pointing this out! I have added the image to the repo and updated the corresponding notebook. Let me know if you or have any other suggestions!
