+++ tests/test_segmenteverygrain.py
@@ -6,51 +6,55 @@ from shapely.geometry import Polygon
 import segmenteverygrain as seg
 import tensorflow as tf
 
+
 class TestGetGrainsFromPatches(unittest.TestCase):
 
     def setUp(self):
         # Create a mock image
         self.image = np.zeros((100, 100, 3), dtype=np.uint8)
-        
+
         # Create a mock Axes object with patches
         self.fig, self.ax = plt.subplots()
         self.patches = [
             MplPolygon(np.array([[10, 10], [20, 10], [20, 20], [10, 20]]), closed=True),
-            MplPolygon(np.array([[30, 30], [40, 30], [40, 40], [30, 40]]), closed=True)
+            MplPolygon(np.array([[30, 30], [40, 30], [40, 40], [30, 40]]), closed=True),
         ]
         for patch in self.patches:
             self.ax.add_patch(patch)
-        plt.axis('equal')
+        plt.axis("equal")
 
     def test_get_grains_from_patches(self):
-        all_grains, rasterized, mask_all = seg.get_grains_from_patches(self.ax, self.image)
-        
+        all_grains, rasterized, mask_all = seg.get_grains_from_patches(
+            self.ax, self.image
+        )
+
         # Check the number of grains
         self.assertEqual(len(all_grains), len(self.patches))
-        
+
         # Check the type of grains
         for grain in all_grains:
             self.assertIsInstance(grain, Polygon)
-        
+
         # Check the rasterized image
         self.assertEqual(rasterized.shape, self.image.shape[:2])
         self.assertTrue(np.any(rasterized > 0))
-        
+
         # Check the mask_all image
         self.assertEqual(mask_all.shape, self.image.shape[:2])
         self.assertTrue(np.any(mask_all == 1))
         self.assertTrue(np.any(mask_all == 2))
-        
+
+
 class TestRasterizeGrains(unittest.TestCase):
 
     def setUp(self):
         # Create a mock image
         self.image = np.zeros((100, 100, 3), dtype=np.uint8)
-        
+
         # Create mock grains
         self.grains = [
             Polygon([(10, 10), (20, 10), (20, 20), (10, 20)]),
-            Polygon([(30, 30), (40, 30), (40, 40), (30, 40)])
+            Polygon([(30, 30), (40, 30), (40, 40), (30, 40)]),
         ]
 
     def test_rasterize_grains_shape(self):
@@ -71,6 +75,7 @@ class TestRasterizeGrains(unittest.TestCase):
         self.assertEqual(rasterized[15, 15], 1)
         self.assertEqual(rasterized[35, 35], 2)
 
+
 class TestPredictImageTile(unittest.TestCase):
     def setUp(self):
         # Create a mock model with a predict method
@@ -90,6 +95,7 @@ class TestPredictImageTile(unittest.TestCase):
         im_tile = np.zeros((256,))
         with self.assertRaises(ValueError):
             seg.predict_image_tile(im_tile, self.model)
+
 
 class TestPredictImage(unittest.TestCase):
     def setUp(self):
@@ -117,12 +123,13 @@ class TestPredictImage(unittest.TestCase):
         with self.assertRaises(ValueError):
             seg.predict_image(big_im, self.model, 256)
 
+
 class TestLabelGrains(unittest.TestCase):
 
     def setUp(self):
         # Create a mock image
         self.image = np.zeros((100, 100, 3), dtype=np.uint8)
-        
+
         # Create a mock prediction
[…]
