+++ segmenteverygrain/segmenteverygrain.py
@@ -22,7 +22,7 @@ from sklearn.model_selection import train_test_split
 from sklearn.cluster import DBSCAN
 import rtree
 import itertools
-from glob import glob 
+from glob import glob
 import os
 from PIL import Image
 import json
@@ -38,6 +38,7 @@ from keras.saving import load_model
 from keras.optimizers import Adam
 
 from segment_anything import SamPredictor
+
 
 def predict_image_tile(im_tile, model):
     """
@@ -55,7 +56,7 @@ def predict_image_tile(im_tile, model):
     im_tile_pred : 3D array
         Predicted tile.
     """
-    
+
     # Check for invalid input
     if not isinstance(im_tile, np.ndarray):
         raise ValueError("Input image tile must be a numpy array.")
@@ -63,10 +64,11 @@ def predict_image_tile(im_tile, model):
         raise ValueError("Input image tile must be a 3D array.")
     if len(im_tile.shape) == 3 and im_tile.shape[2] != 3:
         raise ValueError("3D input image tile must have 3 channels.")
-    im_tile = np.expand_dims(im_tile, axis=0) # add batch dimension
-    im_tile_pred = model.predict(im_tile, verbose=0) # make prediction
-    im_tile_pred = im_tile_pred[0] # remove batch dimension
+    im_tile = np.expand_dims(im_tile, axis=0)  # add batch dimension
+    im_tile_pred = model.predict(im_tile, verbose=0)  # make prediction
+    im_tile_pred = im_tile_pred[0]  # remove batch dimension
     return im_tile_pred
+
 
 def predict_image(image, model, I):
     """
@@ -98,59 +100,79 @@ def predict_image(image, model, I):
     pad_rows = I - np.mod(image.shape[0], I)
     pad_cols = I - np.mod(image.shape[1], I)
     if image.ndim == 2:
-        image = np.stack((image, image, image), axis=-1) # convert to 3 channels
+        image = np.stack((image, image, image), axis=-1)  # convert to 3 channels
     if image.ndim == 3:
         image = np.vstack((image, np.zeros((pad_rows, image.shape[1], image.shape[2]))))
         image = np.hstack((image, np.zeros((image.shape[0], pad_cols, image.shape[2]))))
-    r = int(np.floor(image.shape[0]/I)) # number of rows of image tiles
-    c = int(np.floor(image.shape[1]/I)) # number of columns of image tiles
-    
-    I2 = int(I/2)
+    r = int(np.floor(image.shape[0] / I))  # number of rows of image tiles
+    c = int(np.floor(image.shape[1] / I))  # number of columns of image tiles
+
+    I2 = int(I / 2)
     W = np.hanning(I) * np.hanning(I)[:, np.newaxis]
     Wup = W.copy()
     Wup[:I2, :] = np.tile(np.hanning(I), (I2, 1))
     Wdown = W.copy()
     Wdown[I2:, :] = np.tile(np.hanning(I), (I2, 1))
 
-    image = np.hstack((np.zeros((r*I, I2, 3)), image, np.zeros((r*I, I2, 3)))) # padding on the left and right sides
+    image = np.hstack(
+        (np.zeros((r * I, I2, 3)), image, np.zeros((r * I, I2, 3)))
+    )  # padding on the left and right sides
     image_pred = np.zeros((image.shape[0], image.shape[1], 3))
-    print('segmenting image tiles...')
-    for i in trange(c+1): # rows, no offset
-        for j in range(1,2*r-2): # columns
-            im_tile = image[j*I2:(j+2)*I2, i*I:(i+1)*I, :]/255.0
+    print("segmenting image tiles...")
[…]
