### 2025-04-29 e649b51
Update segmenteverygrain.py
remove unused imports and reorganize standard imports
