### 2025-04-29 be38ab5
Update segmenteverygrain.py
refactor area weighted distribution and add docstring
