### 2025-06-17 a46562a
better package management in colab notebook
