### 2025-06-18 d8250d9
add new interactive interface as segmenteverygrain/interactions submodule
