### 2025-06-18 26aaba7
add tests for new interactive interface
