+++ tests/test_interactions.py
@@ -1,0 +1,275 @@
+# Standard imports
+import math
+import unittest
+# Pip imports
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+import numpy as np
+import pandas as pd
+import shapely
+# Local imports
+import segmenteverygrain as seg
+import segmenteverygrain.interactions as si
+
+
+# Geometry calculations
+def _azimuth(point1: tuple, point2: tuple) -> float:
+    ''' Azimuth between 2 points (interval 0 - 180) '''
+    return np.arctan2(point2[0] - point1[0], point2[1] - point1[1]) / 2
+
+
+def dist(a: tuple, b: tuple) -> float:
+    ''' Distance between points '''
+    return math.hypot(b[0] - a[0], b[1] - a[1])
+
+
+def azimuth(mrr: shapely.Polygon) -> float:
+    ''' Azimuth of minimum_rotated_rectangle '''
+    bbox = list(mrr.exterior.coords)
+    axis1 = dist(bbox[0], bbox[3])
+    axis2 = dist(bbox[0], bbox[1])
+    if axis1 <= axis2:
+        az = _azimuth(bbox[0], bbox[1])
+    else:
+        az = _azimuth(bbox[0], bbox[3])
+    return az
+
+
+# Comparison methods
+def absolute(a, b):
+    return abs(a - b)
+
+
+def relative(a, b):
+    return 100 * abs((a - b) / (a + b)) / 2
+
+
+def absolute_angle(a, b):
+    diff = absolute(a, b)
+    return min(diff, np.pi - diff) * 180. / np.pi
+
+
+class Comparison():
+    def __init__(self, f: object, tolerance: float):
+        # Comparison function
+        self.f = f
+        # Tolerance for comparison
+        self.tolerance = tolerance
+
+
+metrics = {
+    'centroid-0': Comparison(absolute, 2.),
+    'centroid-1': Comparison(absolute, 2.),
+    'area': Comparison(relative, 2.),
+    'perimeter': Comparison(relative, 2.),
+    'orientation': Comparison(absolute_angle, 10.),
+    'major_axis_length': Comparison(relative, 1.),
+    'minor_axis_length': Comparison(relative, 2.),
+    # 'max_intensity-0': Comparison(absolute, 5.),
+    # 'min_intensity-0': Comparison(absolute, 10.),
+    'mean_intensity-0': Comparison(relative, 2.),
+    # 'max_intensity-1': Comparison(absolute, 5.),
+    # 'min_intensity-1': Comparison(absolute, 10.),
+    'mean_intensity-1': Comparison(relative, 2.),
+    # 'max_intensity-2': Comparison(absolute, 5.),
+    # 'min_intensity-2': Comparison(absolute, 10.),
+    'mean_intensity-2': Comparison(relative, 2.)
+}
+
+
+class TestPolygonMeasurement(unittest.TestCase):
+    ''' Test polygon measurement functions. '''
+
+    @classmethod
+    def setUpClass(cls):
+        # Create image with random color
+        cls.rgb = np.random.randint(0, 255, size=3, dtype=np.uint8)
+        cls.image = np.zeros((100, 100, 3), dtype=np.uint8)
+        cls.image[:, :] = cls.rgb
+        # Create polygons
+        cls.polygons = [shapely.Polygon(coords) for coords in [
+            [[10, 10], [25, 15], [25, 25], [15, 25]],
+            [[30, 30], [40, 30], [40, 40], [30, 40]]
+        ]]
+
+    def test_measure_polygon(self):
+        for poly in self.polygons:
+            data = si.measure_polygon(poly)
+            # Check the centroid
+            self.assertEqual(data['centroid'], poly.centroid.xy[::-1])
[…]
