### 2025-06-18 e307213
use interactive_edit example for tests to work with pull request
