### 2025-06-18 4cf5b28
Move predict_from_prompts to interactions.py to keep segmenteverygrain.py unchanged
