### 2025-06-23 zsylvester/segmenteverygrain#33 @pr4deepr
recommendations for installation
Hi
   Some recommendations on documentations for installation

- specifying the recommended python version if using `pip install`
- Using Miniforge or Miniconda instead of anaconda due to licensing restrictions. Also, when I used `environment.yml` with `defaults` channel enabled,  I couldn't download packages as my institute has blocked anaconda and defaults channel due to licensing restrictions.
- Emphasize that you need to clone the repo in order to run the software, especially due to the model files.

You already have some of this information. Maybe making it as dot points will clarify it.

https://github.com/openjournals/joss-reviews/issues/7953

Cheers
Pradeep
