### 2025-06-23 zsylvester/segmenteverygrain#34 @pr4deepr
SAM model installation
Hi
   Just like in the google colab notebook, you can have some code to download the `SAM` model in the jupyter notebook.
This will automate the process rather than a manual download.

This worked for me:

```
import os
if not os.path.exists("./models/sam_vit_h_4b8939.pth"):
    import urllib.request
    url = "https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth"
    urllib.request.urlretrieve(url, "./models/sam_vit_h_4b8939.pth")
```

Cheers
Pradeep

https://github.com/openjournals/joss-reviews/issues/7953
