### 2025-06-23 zsylvester/segmenteverygrain#35 @pr4deepr
About section (minor)
A minor suggestion:

Adding the documentation link in your `About` section on the top right will make it easier to navigate to documentation when visiting the github  page.

![Image](https://github.com/user-attachments/assets/1e1ccbdf-4fa5-4341-93e5-322f52fc3d5a)

https://github.com/openjournals/joss-reviews/issues/7953
