### 2025-06-23 zsylvester/segmenteverygrain#36 @pr4deepr
image sources and info
Hi
   In the paper, you use a couple of images and show some graphs.
I couldn't find the same image on your repo. I could see some references in the manuscript.
- How do I get access to these images if I would like to reproduce your results?

On another note, I do not see any information on the image source/s under [examples](https://github.com/zsylvester/segmenteverygrain/tree/main/examples) folder. Could you add a README, indicating where each of the images were from?
Ideally, you would want to add some information on acquisition details, but you could point to an existing publication if that helps.

Cheers
Pradeep

https://github.com/openjournals/joss-reviews/issues/7953
