### 2025-06-23 zsylvester/segmenteverygrain#37 @pr4deepr
Documentation clarity
Hi
   The documentation is looking good. I would make some suggestions to improve clarity, otherwise the details get lost in the text.
Doc website:

- In the `Getting Started` section, you could add a link to the `[Segment_every_grain.ipynb](https://github.com/zsylvester/segmenteverygrain/blob/main/segmenteverygrain/Segment_every_grain.ipynb) notebook at the very start indicating that there is a much more comprehensive tutorial in the notebook.
- You can consider integrating your notebook into the doc website and/OR perhaps adding an `nbviewer` option to render your notebooks.

Cheers
Pradeep

https://github.com/openjournals/joss-reviews/issues/7953
