### 2025-06-23 zsylvester/segmenteverygrain#38 @pr4deepr
running time
It is worth adding some information about time it takes to run segmentation.
You can use the example datasets and mention time on google colab.
i..e, With an Nvidia GPU ABC model
- for images of XY resolution of this size A w it took 3 minutes
- for large image of XY resolution of size B w it took 30 minutes

Its not essential for the JOSS review.

https://github.com/openjournals/joss-reviews/issues/7953
