### 2025-06-23 #7953 @pr4deepr
Hi
    I've been enjoying reviewing this software. I've been raising issues on the github repo.
I'm quite happy with the software and the manuscript. I have a few questions/suggestions:

- How is your software different than SediNet? SediNet seems to give a grain size profile as well and claims to adapt well to wider spectrum of grain types/sizes.
- Is your software easier to use and adapt compared to others? If so, its worth mentioning that in the statement of need.
- Your code is currently one big `.py` file. For maintainability, I would recommend splitting it into different files. This is not essential for this review though.
- Was the original U-Net model trained using the same code as [create_and_train_model](https://github.com/zsylvester/segmenteverygrain/blob/a46562aa422fb7fb46e4d39399139ae244987f0f/segmenteverygrain/segmenteverygrain.py#L2014)? If so, were these from the sedinet project? It just wasn't clear which images were used exactly. It will be great to add some detail on how the base U-Net was trained, i.e., exact dataset, loss curves and any settings used etc...  This enables reproducibility. This can be added to the repo and doesn't need to be in the manuscript.
- Could you add some recommended hardware specs?

Cheers
Pradeep
