### 2025-06-30 e60adce
add details about running time
