### 2025-06-30 4503943
tweak installation instructions in Readme'
