+++ environment.yml
@@ -2,7 +2,6 @@ name: segmenteverygrain
 
 channels:
   - conda-forge
-  - defaults
 
 dependencies:
   - python=3.9
