### 2025-06-30 zsylvester/segmenteverygrain#33 @zsylvester
Hi @pr4deepr -- I have updated the installation instructions so that they don't recommend using Anaconda. Also, I removed the 'defaults' channel from the YML files.
