### 2025-06-30 a69d8c0
minor tweaks to Readme
