### 2025-06-30 zsylvester/segmenteverygrain#34 @zsylvester
Thanks @pr4deepr, that's a great suggestion. I have added the code snippet to the notebook.
