### 2025-06-30 8d53ce9
add readme files to example images
