+++ examples/mair_et_al_L2_DJI_0382/readme.md
@@ -1,0 +1,5 @@
+# Example Image from the Swiss Alps
+
+The image in this folder comes from the data collected by Mair et al. (2022) from gravel bars of rivers in the Swiss Alps, using drones and structure-from-motion software. See the paper below for more details. This image is used in the `segmenteverygrain` example notebook to illustrate how the segmentation can be run on large images.
+
+Mair, D., Do Prado, A.H., Garefalakis, P., Lechmann, A., Whittaker, A., and Schlunegger, F., 2022, Grain size of fluvial gravel bars from close-range UAV imagery – uncertainty in segmentation-based data: Earth Surface Dynamics, v. 10, p. 953–973, doi:10.5194/esurf-10-953-2022.
