+++ examples/miocene_sandstone/readme.md
@@ -1,0 +1,5 @@
+# Thin section example image - Sandstone from the Gulf of Mexico
+
+This image is a small subset of the original data that was collected from a thin section of a Miocene sandstone from the subsurface of the offshore Gulf of Mexico. See the refernec below for more details and more data.
+
+Prodanovic, M., Milliken, K. L., & Mehmani, A. (2019). Offshore miocene sandstone thin section (Folk McBride collection). https://www.digitalrocksportal.org/projects/244; Digital Rocks Portal. https://doi.org/10.17612/rn6g-fq64
