### 2025-06-30 zsylvester/segmenteverygrain#36 @zsylvester
Thanks @pr4deepr - I have added the two larger images in the paper to the 'examples' folder. I have also added short README files to the folders that briefly explain what the images are and point to the original references (if any).
