+++ _sources/getting_started.rst.txt
@@ -4,7 +4,14 @@ Getting started
 .. toctree::
    :caption: Getting started
 
-To load the U-Net model:
+The best way to use the `segmenteverygrain` package is to run the `Segment_every_grain.ipynb <https://github.com/zsylvester/segmenteverygrain/blob/main/segmenteverygrain/Segment_every_grain.ipynb>`_ notebook.
+
+The notebook goes through the steps of loading the models, running the segmentation, interactively updating the result, and saving the grain data and the mask. The text below summarizes the steps that you need to take to run the segmentation.
+
+Loading the models
+~~~~~~~~~~~~~~~~~~
+
+To load the U-Net model, you can use the 'load_model' function from Keras. The U-Net model is saved in the 'seg_model.keras' file.
 
 .. code-block:: python
 
@@ -14,28 +21,56 @@ To load the U-Net model:
 
 This assumes that you are using Keras 3 and 'seg_model.keras' was saved using Keras 3. Older models created with a ``segmenteverygrain`` version that was based on Keras 2 do not work with with the latest version of the package.
 
+The Segment Anything model can be downloaded from this `link <https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth>`_. You can also download it programmatically:
+
+.. code-block:: python
+
+   import os
+   import urllib.request
+   
+   if not os.path.exists("./models/sam_vit_h_4b8939.pth"):
+       url = "https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth"
+       urllib.request.urlretrieve(url, "./models/sam_vit_h_4b8939.pth")
+
+Running the segmentation
+~~~~~~~~~~~~~~~~~~~~~~~~
+
 To run the U-Net segmentation on an image and label the grains in the U-Net output:
 
 .. code-block:: python
 
+   import numpy as np
+   import matplotlib.pyplot as plt
+   from keras.utils import load_img
+   
+   # Load your image
+   fname = "path/to/your/image.jpg"
+   image = np.array(load_img(fname))
+   
+   # Run U-Net prediction
    image_pred = seg.predict_image(image, model, I=256)
    labels, coords = seg.label_grains(image, image_pred, dbs_max_dist=20.0)
     
 The input image should not be much larger than ~2000x3000 pixels, in part to avoid long running times; it is supposed to be a numpy array with 3 channels (RGB).
 Grains should be well defined in the image and not too small (e.g., only a few pixels in size).
+
+Quality control of U-Net prediction
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
 The U-Net prediction should be QC-d before running the SAM segmentation:
 
 .. code-block:: python
 
    plt.figure(figsize=(15,10))
-   plt.imshow(big_im_pred)
+   plt.imshow(image_pred)
    plt.scatter(np.array(coords)[:,0], np.array(coords)[:,1], c='k')
    plt.xticks([])
    plt.yticks([])
 
-If the U-Net segmentation is of low quality, the base model can be (and should be) finetuned using the steps outlined :ref:`below<Finetuning the U-Net model>`.
+The black dots in the figure represent the SAM prompts that will be used for grain segmentation. If the U-Net segmentation is of low quality, the base model can be (and should be) finetuned using the steps outlined :ref:`below<Finetuning the U-Net model>`.
 
-The Segment Anything model can be downloaded from this `link <https://dl.fbaipublicfiles.com/segment_anything/sam_vit_h_4b8939.pth>`_.
+SAM segmentation
+~~~~~~~~~~~~~~~~
 
 Here is an example showing how to run the SAM segmentation on an image, using the outputs from the U-Net model:
 
[…]
