### 2025-07-01 zsylvester/segmenteverygrain#37 @zsylvester
Hi @pr4deepr - great suggestions! I have now updated the '[Getting started](https://zsylvester.github.io/segmenteverygrain/getting_started.html)' section of the documentation so that it largely parallels the contents of the notebook. I have also added a reference / link to the notebook at the beginning of the page.
