+++ README.md
@@ -1,6 +1,6 @@
 # segmenteverygrain
 
-<img src="docs/gravel_example_mask.png" width="600">
+<img src="docs/_static/images/gravel_example_mask.png" width="600">
 
 [![Tests](https://github.com/zsylvester/segmenteverygrain/actions/workflows/ci.yaml/badge.svg)](https://github.com/zsylvester/segmenteverygrain/actions/workflows/ci.yaml)
 [![License](https://img.shields.io/badge/License-Apache_2.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
