### 2025-07-03 #7953 @pr4deepr
Thanks a lot @zsylvester. This is great and I'm happy with the responses!
Two comments:
1. The link to [Train_Unet_model.ipynb](https://github.com/zsylvester/segmenteverygrain/blob/main/segmenteverygrain/Train_Unet_model.ipynb) at this documentation[ link ](https://zsylvester.github.io/segmenteverygrain/getting_started.html#training-the-u-net-model-from-scratch)does not work. However, your link in the post above does.
2. Could you regenerate the updated pdf ?? Not sure if I can do it.

@anastassiavybornova , I've finished my review and the authors have a done great job with addressing the comments, as well as with the software!
