### 2025-07-03 a26b393
minor tweaks
