### 2025-07-03 #7953 @zsylvester
Thanks @pr4deepr - I fixed the notebook link in the documentation (and found another link that was broken as well).

You can download the updated paper PDF at the bottom of [this page](https://github.com/zsylvester/segmenteverygrain/actions/runs/16039850899).
