### 2025-08-04 81ea65d
minor tweaks to paper and bibiliography
