+++ paper/paper.bib
@@ -388,7 +388,7 @@
 }
 
 @software{Gillies:2013,
-	author =    {Sean Gillies et al.},
+	author =    {Gillies, Sean, et al.},
 	organization = {Mapbox},
 	title =     {Rasterio: geospatial raster I/O for {Python} programmers},
 	year =      {2013},
@@ -396,7 +396,7 @@
 }
 
 @Misc{Gillies:2007,
-	author = {Sean Gillies et al.},
+	author = {Gillies, Sean, et al.},
 	organization = {toblerity.org},
 	title = {Shapely: manipulation and analysis of geometric objects},
 	year =  {2007},
