### 2025-08-04 9727c0b
minor tweaks to paper and bibiliography
