+++ paper/paper.bib
@@ -387,20 +387,21 @@
 	pages = {e453},
 }
 
-@software{Gillies:2013,
-	author =    {Gillies, Sean, et al.},
-	organization = {Mapbox},
-	title =     {Rasterio: geospatial raster I/O for {Python} programmers},
-	year =      {2013},
-	url = "https://github.com/rasterio/rasterio"
+@software{Gillies:2019,
+  author =    {Gillies, Sean and others},
+  organization = {Mapbox},
+  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
+  year =      {2013--},
+  url = {https://github.com/rasterio/rasterio}
 }
 
-@Misc{Gillies:2007,
-	author = {Gillies, Sean, et al.},
-	organization = {toblerity.org},
-	title = {Shapely: manipulation and analysis of geometric objects},
-	year =  {2007},
-	url = "https://github.com/Toblerity/Shapely"
+@software{Gillies:2025,
+	author = {Gillies, S. and van der Wel, C. and Van den Bossche, J. and Taves, M. W. and Arnott, J. and Ward, B. C. and others},
+	publisher = {Zenodo},
+	title = {Shapely (2.1.1)},
+	year =  {2025},
+	url = {https://doi.org/10.5281/zenodo.15463269},
+	doi = {10.5281/zenodo.15463269},
 }
 
 @inproceedings{Hagberg:2008,
