### 2025-08-04 6fda84c
minor tweaks to bibiliography
