+++ paper/paper.bib
@@ -388,7 +388,7 @@
 }
 
 @software{Gillies:2019,
-  author =    {Gillies, Sean and others},
+  author =    {Gillies, S. and others},
   organization = {Mapbox},
   title =     {Rasterio: geospatial raster I/O for {Python} programmers},
   year =      {2013--},
