### 2025-08-04 #7953 @zsylvester
Hi @anastassiavybornova - thank you for handling this manuscript and guiding the review!

I have now added the sources for the photos in the paper in the figure captions and reformatted the references for shapely and rasterio.

I have also created a tagged release (version 0.2.3) and linked the repo to a [Zenodo archive](https://doi.org/10.5281/zenodo.16740743); the DOI is: 10.5281/zenodo.16740743.

Let me know if there is anything else I should do.
