### 2025-08-05 48d3585
minor tweaks to paper
