### 2025-08-05 1cac9b2
minor tweaks to paper
