### 2025-08-05 fd3ae85
fix some issues with bibliography
