+++ paper/paper.bib
@@ -1,37 +1,4 @@
 
-@article{Paola:1996,
-	title = {Palaeohydraulics revisited: palaeoslope estimation in coarse-grained braided rivers},
-	volume = {8},
-	issn = {0950-091X, 1365-2117},
-	shorttitle = {Palaeohydraulics revisited},
-	url = {https://onlinelibrary.wiley.com/doi/10.1046/j.1365-2117.1996.00253.x},
-	doi = {10.1046/j.1365-2117.1996.00253.x},
-	language = {en},
-	number = {3},
-	urldate = {2024-12-16},
-	journal = {Basin Research},
-	author = {Paola, Chris and Mohrig, David},
-	month = sep,
-	year = {1996},
-	pages = {243--254},
-}
-
-@article{Buscombe:2013,
-	title = {Transferable wavelet method for grain-size distribution from images of sediment surfaces and thin sections, and other natural granular patterns},
-	volume = {60},
-	issn = {0037-0746, 1365-3091},
-	url = {https://onlinelibrary.wiley.com/doi/10.1111/sed.12049},
-	doi = {10.1111/sed.12049},
-	language = {en},
-	number = {7},
-	urldate = {2024-12-16},
-	journal = {Sedimentology},
-	author = {Buscombe, Daniel},
-	month = dec,
-	year = {2013},
-	pages = {1709--1732},
-}
-
 @article{Buscombe:2020,
 	title = {SediNet: a configurable deep learning model for mixed qualitative and quantitative optical granulometry},
 	volume = {45},
@@ -65,25 +32,8 @@
 	pages = {1234--1250},
 }
 
-@article{Parker:2007,
-	title = {Physical basis for quasi-universal relations describing bankfull hydraulic geometry of single‐thread gravel bed rivers},
-	volume = {112},
-	copyright = {http://onlinelibrary.wiley.com/termsAndConditions\#vor},
-	issn = {0148-0227},
-	url = {https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2006JF000549},
-	doi = {10.1029/2006JF000549},
-	language = {en},
-	number = {F4},
-	urldate = {2024-12-16},
-	journal = {Journal of Geophysical Research: Earth Surface},
-	author = {Parker, Gary and Wilcock, Peter R. and Paola, Chris and Dietrich, William E. and Pitlick, John},
-	month = dec,
-	year = {2007},
-	pages = {2006JF000549},
-}
-
 @article{Kirillov:2023,
-	title = {Segment Anything},
+	title = {Segment {A}nything},
 	url = {http://arxiv.org/abs/2304.02643},
 	doi = {10.48550/arXiv.2304.02643},
 	language = {en},
@@ -131,7 +81,7 @@
 }
 
 @article{Purinton:2021,
-	title = {Tracking Downstream Variability in Large Grain-Size Distributions in the South-Central Andes},
+	title = {Tracking {D}ownstream {V}ariability in {L}arge {G}rain-{S}ize {D}istributions in the {S}outh-{C}entral {A}ndes},
 	volume = {126},
 	issn = {2169-9003, 2169-9011},
 	url = {https://agupubs.onlinelibrary.wiley.com/doi/10.1029/2021JF006260},
@@ -144,21 +94,6 @@
 	month = aug,
 	year = {2021},
 	pages = {e2021JF006260},
-}
-
-@article{Reynolds:2024,
-	title = {Grain size from source to sink - modern and ancient fining rates},
-	volume = {250},
-	issn = {00128252},
-	url = {https://linkinghub.elsevier.com/retrieve/pii/S0012825224000266},
[…]
