### 2025-08-07 #7953 @zsylvester
Hi @anastassiavybornova - do I need to create another release / archive that includes the minor changes in the paper that were implemented after version 0.2.3 was released?
