### 2025-08-08 #7953 @anastassiavybornova
@zsylvester hi again! so :) for the minor changes in the paper - it is not necessary to create a new release (would only be necessary in case changes to the *software* occur; and since 0.2.3. is now the latest version of the software that includes the changes to the *software* implemented during the review process that's fine)

and for the archive at Zenodo - the author list on zenodo should be exactly the same as the author list of the paper, could you please update this and ping me here once it's done
