### 2025-08-08 #7953 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
